requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
catboost
xgboost
dill
flask
streamlit
pdfkit
fpdf
joblib
streamlit-lottie
google-generativeai 
python-dotenv
#-e .